atlos_v5/requirements.txt
# ATLOS v5.0 - Dependencies
# Installation: pip install -r requirements.txt

# Core dependencies
pyyaml>=6.0
cryptography>=41.0.0
requests>=2.31.0
rich>=13.0.0
click>=8.1.0

# Network scanning
scapy>=2.5.0
python-nmap>=0.7.1

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# API REST
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.0

# LDAP enumeration
ldap3>=2.9.1

# HTTP/HTTPS handling
urllib3>=2.0.0
certifi>=2023.7.22

# Data processing
pandas>=2.1.0
numpy>=1.24.0

# Reporting
jinja2>=3.1.0
weasyprint>=60.0
plotly>=5.17.0

# Async support
aiohttp>=3.9.0
aiofiles>=23.2.1

# Testing (development)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development tools
black>=23.9.0
flake8>=6.1.0
mypy>=1.6.0

# Security tools
bandit>=1.7.5
safety>=2.3.0

# Documentation
sphinx>=7.2.0
sphinx-rtd-theme>=1.3.0

# Optional: Cloud providers
# boto3>=1.29.0  # AWS
# azure-identity>=1.15.0  # Azure
# google-cloud-compute>=1.14.0  # GCP

# Optional: Container security
# docker>=6.1.0
# kubernetes>=28.1.0

# Optional: Machine learning for threat detection
# scikit-learn>=1.3.0
# tensorflow>=2.14.0
